requirements.txt
fastapi==0.92.0
orjson==3.8.7
python-logstash==0.4.8
uvicorn==0.20.0
pymongo==4.3.3
motor==3.1.1
gunicorn==20.1.0
python-dotenv==1.0.0
fastapi-admin==1.0.4
django==4.1.7
aiohttp==3.8.3
requests==2.28.2
schedule==1.1.0
SQLAlchemy==2.0.6
pika==1.3.1
psycopg2-binary==2.9.5
flask==2.2.3
Flask-Migrate==4.0.2
motor==3.1.1
django-better-admin-arrayfield==1.4.2
mailhog==0.1.0